Bug fixed that produced duplicated 'id'
There was bug in function util.find_next_id
When we had id's:0, 1, 3, 4 it added 2 at the end because after 1 it found empty space.
And we got 0, 1, 3, 4, 2
and during next call it also added 2 because still after 1 was 3 (missed 2) nevertheless it was at the end
So we got two 2's at the end (and so on). It produced big mess.
Bug was fixed - now before iterating on dictionaries list, the list is sort by dictionaries id values.