Update database file.
Now in comment table question_id excludes answer_id and vice versa
what means that it is not possible to add comment that is tight both to quesion and answer.

Also in users table added column inactive (boolean) that is set default as False.
Now it is not obligatory to delete user, we can inactivate them.